+++ packages/http/src/errors/errors.ts
@@ -1,2 +1,3 @@
 export * from "./http-client-request.error";
 export * from "./http-client-response-redirect.error";
+export * from "./payload-too-large.error";